Function quadratic::jacobi

pub fn jacobi(a_: i32, n_: i32) -> i8

Returns the value of the Jacobi symbol for (a \ n) Where a is an integer and n is an odd positive integer

Examples

// compute the Jacobi symbol of (2 \ 5)
let symb = quadratic::jacobi(2, 5);