Crate quadratic

Helper functions related to quadratic reciprocity.

This library calculates the Legendre and Jacobi symbol using the standard algorithm.

Since the Jacobi symbol is a generalization of the Legendre symbol, only a jacobi function is provided.

Functions

jacobi
Returns the value of the Jacobi symbol for (a \ n) Where a is an integer and n is an odd positive integer